+++ nts.uk/uk.pr/pr.ctx/pr.core/nts.uk.ctx.pr.core.app/src/main/java/nts/uk/ctx/pr/core/app/find/wageprovision/statementitem/ItemRangeSetDto.java
@@ -21,7 +21,7 @@ public class ItemRangeSetDto {
 	/**
 	 * エラー上限値金額
 	 */
-	private BigDecimal errorUpperRangeValueAmount;
+	private Long errorUpperRangeValueAmount;
 
 	/**
 	 * エラー上限値時間
@@ -41,7 +41,7 @@ public class ItemRangeSetDto {
 	/**
 	 * エラー上限値金額
 	 */
-	private BigDecimal errorLowerRangeValueAmount;
+	private Long errorLowerRangeValueAmount;
 
 	/**
 	 * エラー上限値時間
@@ -61,7 +61,7 @@ public class ItemRangeSetDto {
 	/**
 	 * アラーム上限値金額
 	 */
-	private BigDecimal alarmUpperRangeValueAmount;
+	private Long alarmUpperRangeValueAmount;
 
 	/**
 	 * アラーム上限値時間
@@ -81,7 +81,7 @@ public class ItemRangeSetDto {
 	/**
 	 * アラーム上限値金額
 	 */
-	private BigDecimal alarmLowerRangeValueAmount;
+	private Long alarmLowerRangeValueAmount;
 
 	/**
 	 * アラーム上限値時間
